Revert embeddings Don't remake made vocabs

+++ code/constrained/src/main.py
@@ -171,31 +171,43 @@ def main():
             if is_train:
                 train_dataloader, val_dataloader = load_data(config, logger)
 
-                logger.debug("Creating Vocab...")
+                if not os.path.exists(vocab1_path):
+                    logger.debug("Creating Vocab 1...")
 
-                voc1 = Voc1()
-                voc1.create_vocab_dict(config, train_dataloader)
+                    voc1 = Voc1()
+                    voc1.create_vocab_dict(config, train_dataloader)
 
-                # To Do : Remove Later
-                voc1.add_to_vocab_dict(config, val_dataloader)
+                    # To Do : Remove Later
+                    voc1.add_to_vocab_dict(config, val_dataloader)
 
-                voc2 = Voc2(config)
-                voc2.create_vocab_dict(config, train_dataloader)
+                    with open(vocab1_path, "wb") as f:
+                        pickle.dump(voc1, f, protocol=pickle.HIGHEST_PROTOCOL)
 
-                # To Do : Remove Later
-                voc2.add_to_vocab_dict(config, val_dataloader)
+                    logger.info("Vocab saved at {}".format(vocab1_path))
+                else:
+                    with open(vocab1_path, "rb") as f:
+                        voc1 = pickle.load(f)
+
+                if not os.path.exists(vocab2_path):
+                    logger.debug("Creating Vocab 2...")
+
+                    voc2 = Voc2(config)
+                    voc2.create_vocab_dict(config, train_dataloader)
+
+                    # To Do : Remove Later
+                    voc2.add_to_vocab_dict(config, val_dataloader)
+
+                    with open(vocab2_path, "wb") as f:
+                        pickle.dump(voc2, f, protocol=pickle.HIGHEST_PROTOCOL)
+
+                    logger.info("Vocab saved at {}".format(vocab2_path))
+                else:
+                    with open(vocab2_path, "rb") as f:
+                        voc2 = pickle.load(f)
 
                 logger.info(
                     "Vocab Created with number of words : {}".format(voc1.nwords)
                 )
-
-                with open(vocab1_path, "wb") as f:
-                    pickle.dump(voc1, f, protocol=pickle.HIGHEST_PROTOCOL)
-                with open(vocab2_path, "wb") as f:
-                    pickle.dump(voc2, f, protocol=pickle.HIGHEST_PROTOCOL)
-
-                logger.info("Vocab saved at {}".format(vocab1_path))
-
             else:
                 test_dataloader = load_data(config, logger)
                 logger.info("Loading Vocab File...")
@@ -359,29 +371,41 @@ def main():
         if is_train:
             train_dataloader, val_dataloader = load_data(config, logger)
 
-            logger.debug("Creating Vocab...")
+            if not os.path.exists(vocab1_path):
+                logger.debug("Creating Vocab 1...")
 
-            voc1 = Voc1()
-            voc1.create_vocab_dict(config, train_dataloader)
+                voc1 = Voc1()
+                voc1.create_vocab_dict(config, train_dataloader)
 
-            # To Do : Remove Later
-            voc1.add_to_vocab_dict(config, val_dataloader)
+                # To Do : Remove Later
+                voc1.add_to_vocab_dict(config, val_dataloader)
 
-            voc2 = Voc2(config)
-            voc2.create_vocab_dict(config, train_dataloader)
+                with open(vocab1_path, "wb") as f:
+                    pickle.dump(voc1, f, protocol=pickle.HIGHEST_PROTOCOL)
 
-            # To Do : Remove Later
-            voc2.add_to_vocab_dict(config, val_dataloader)
+                logger.info("Vocab saved at {}".format(vocab1_path))
+            else:
+                with open(vocab1_path, "rb") as f:
+                    voc1 = pickle.load(f)
+
+            if not os.path.exists(vocab2_path):
+                logger.debug("Creating Vocab 2...")
+
+                voc2 = Voc2(config)
+                voc2.create_vocab_dict(config, train_dataloader)
+
+                # To Do : Remove Later
+                voc2.add_to_vocab_dict(config, val_dataloader)
+
+                with open(vocab2_path, "wb") as f:
+                    pickle.dump(voc2, f, protocol=pickle.HIGHEST_PROTOCOL)
+
+                logger.info("Vocab saved at {}".format(vocab2_path))
+            else:
+                with open(vocab2_path, "rb") as f:
+                    voc2 = pickle.load(f)
 
             logger.info("Vocab Created with number of words : {}".format(voc1.nwords))
-
-            with open(vocab1_path, "wb") as f:
-                pickle.dump(voc1, f, protocol=pickle.HIGHEST_PROTOCOL)
-            with open(vocab2_path, "wb") as f:
-                pickle.dump(voc2, f, protocol=pickle.HIGHEST_PROTOCOL)
-
-            logger.info("Vocab saved at {}".format(vocab1_path))
-
         else:
             test_dataloader = load_data(config, logger)
             logger.info("Loading Vocab File...")
